re-organize controls per design spec
